DONE-nmpt2019 and nmpt-2043

+++ NMP-Portal/NMP-Portal/Models/AutumnCropNitrogenUptakeDetail.cs
@@ -6,6 +6,6 @@
         public string FieldName { get; set; } = string.Empty;
         public int CropTypeId { get; set; }
         public string? CropTypeName { get; set; } = string.Empty;
-        public decimal AutumnCropNitrogenUptake { get; set; }
+        public int AutumnCropNitrogenUptake { get; set; }
     }
 }
